move remote browsing code to Addon\Remote and Layout\Remote
